requirements.txt
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.16
django-allauth==0.42.0
django-currentuser==0.5.1
django-rest-auth==0.9.5
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
flake8==3.8.4
idna==2.10
importlib-metadata==2.0.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
oauthlib==3.1.0
Pillow==7.2.0
pip==20.2.3
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.6.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
setuptools==50.0.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wheel==0.35.1
wrapt==1.12.1
zipp==1.2.0
